Move simple G Suite components out of my-site module
