+++ client/components/gsuite/gsuite-features/compact.jsx
@@ -7,7 +7,7 @@ import React from 'react';
 /**
  * Internal dependencies
  */
-import GSuiteFeatures from 'my-sites/email/gsuite-features';
+import GSuiteFeatures from 'components/gsuite/gsuite-features';
 
 export default function GSuiteCompactFeatures( props ) {
 	return <GSuiteFeatures { ...props } compact />;
